Allow skipping preflight tests using PR labels
By adding a label to a PR in format tests/skip/{test_name} admins can
skip certain tests and make it optional.

JIRA: ISV-4316
